python-service/requirements.txt
# Core Processing Libraries
docling==2.54.0
pandas==2.3.2
openpyxl==3.1.5
pyarrow==16.1.0

# File Detection & Validation
python-magic==0.4.27
charset-normalizer==3.3.2

# Data Validation & Models
pydantic==2.8.2
jsonschema==4.23.0

# API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# CLI & Utils (Optional)
typer==0.12.5
rich==13.9.3
tabulate==0.9.0

# HTTP Client for Supabase Integration
httpx==0.25.2
supabase==2.0.2

# Security & Environment
python-dotenv==1.0.0
bcrypt==4.1.2

# Optional Performance Analysis
polars==1.9.0